requirements.txt
annotated-types==0.7.0 ; python_version >= "3.11" and python_version < "4.0"
anyio==4.8.0 ; python_version >= "3.11" and python_version < "4.0"
async-timeout==5.0.1 ; python_version >= "3.11" and python_full_version < "3.11.3"
bidict==0.23.1 ; python_version >= "3.11" and python_version < "4.0"
blinker==1.9.0 ; python_version >= "3.11" and python_version < "4.0"
cachetools==5.5.0 ; python_version >= "3.11" and python_version < "4.0"
certifi==2024.12.14 ; python_version >= "3.11" and python_version < "4.0"
cffi==1.17.1 ; python_version >= "3.11" and python_version < "4.0" and platform_python_implementation != "PyPy"
charset-normalizer==3.4.1 ; python_version >= "3.11" and python_version < "4.0"
click==8.1.8 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and platform_system == "Windows"
cryptography==44.0.0 ; python_version >= "3.11" and python_version < "4.0"
cssutils==1.0.2 ; python_version >= "3.11" and python_version < "4.0"
diff-match-patch==20181111 ; python_version >= "3.11" and python_version < "4.0"
distro==1.9.0 ; python_version >= "3.11" and python_version < "4.0"
dnspython==2.7.0 ; python_version >= "3.11" and python_version < "4.0"
email-validator==2.2.0 ; python_version >= "3.11" and python_version < "4.0"
et-xmlfile==2.0.0 ; python_version >= "3.11" and python_version < "4.0"
flask-login==0.6.3 ; python_version >= "3.11" and python_version < "4.0"
flask-moment==1.0.6 ; python_version >= "3.11" and python_version < "4.0"
flask-socketio==5.5.1 ; python_version >= "3.11" and python_version < "4.0"
flask-sqlalchemy==3.1.1 ; python_version >= "3.11" and python_version < "4.0"
flask-wtf==1.2.2 ; python_version >= "3.11" and python_version < "4.0"
flask==3.1.0 ; python_version >= "3.11" and python_version < "4.0"
google-ai-generativelanguage==0.6.10 ; python_version >= "3.11" and python_version < "4.0"
google-api-core==2.24.0 ; python_version >= "3.11" and python_version < "4.0"
google-api-core[grpc]==2.24.0 ; python_version >= "3.11" and python_version < "4.0"
google-api-python-client==2.158.0 ; python_version >= "3.11" and python_version < "4.0"
google-auth-httplib2==0.2.0 ; python_version >= "3.11" and python_version < "4.0"
google-auth==2.37.0 ; python_version >= "3.11" and python_version < "4.0"
google-cloud-core==2.4.1 ; python_version >= "3.11" and python_version < "4.0"
google-cloud-secret-manager==2.22.0 ; python_version >= "3.11" and python_version < "4.0"
google-cloud-storage==2.19.0 ; python_version >= "3.11" and python_version < "4.0"
google-cloud==0.34.0 ; python_version >= "3.11" and python_version < "4.0"
google-crc32c==1.6.0 ; python_version >= "3.11" and python_version < "4.0"
google-generativeai==0.8.3 ; python_version >= "3.11" and python_version < "4.0"
google-resumable-media==2.7.2 ; python_version >= "3.11" and python_version < "4.0"
googleapis-common-protos==1.66.0 ; python_version >= "3.11" and python_version < "4.0"
googleapis-common-protos[grpc]==1.66.0 ; python_version >= "3.11" and python_version < "4.0"
greenlet==3.1.1 ; python_version < "3.14" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_version >= "3.11"
grpc-google-iam-v1==0.14.0 ; python_version >= "3.11" and python_version < "4.0"
grpcio-status==1.69.0 ; python_version >= "3.11" and python_version < "4.0"
grpcio==1.69.0 ; python_version >= "3.11" and python_version < "4.0"
gunicorn==23.0.0 ; python_version >= "3.11" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.11" and python_version < "4.0"
hashids==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
httpcore==1.0.7 ; python_version >= "3.11" and python_version < "4.0"
httplib2==0.22.0 ; python_version >= "3.11" and python_version < "4.0"
httpx==0.28.1 ; python_version >= "3.11" and python_version < "4.0"
idna==3.10 ; python_version >= "3.11" and python_version < "4.0"
itsdangerous==2.2.0 ; python_version >= "3.11" and python_version < "4.0"
jinja2==3.1.5 ; python_version >= "3.11" and python_version < "4.0"
jiter==0.8.2 ; python_version >= "3.11" and python_version < "4.0"
lxml==4.9.4 ; python_version >= "3.11" and python_version < "4.0"
markdown==3.7 ; python_version >= "3.11" and python_version < "4.0"
markupsafe==3.0.2 ; python_version >= "3.11" and python_version < "4.0"
openai==1.59.6 ; python_version >= "3.11" and python_version < "4.0"
openpyxl==3.1.5 ; python_version >= "3.11" and python_version < "4.0"
packaging==24.2 ; python_version >= "3.11" and python_version < "4.0"
passlib==1.7.4 ; python_version >= "3.11" and python_version < "4.0"
pdfkit==1.0.0 ; python_version >= "3.11" and python_version < "4.0"
proto-plus==1.25.0 ; python_version >= "3.11" and python_version < "4.0"
protobuf==5.29.3 ; python_version >= "3.11" and python_version < "4.0"
psycopg2-binary==2.9.10 ; python_version >= "3.11" and python_version < "4.0"
pyasn1-modules==0.4.1 ; python_version >= "3.11" and python_version < "4.0"
pyasn1==0.6.1 ; python_version >= "3.11" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.11" and python_version < "4.0" and platform_python_implementation != "PyPy"
pydantic-core==2.27.2 ; python_version >= "3.11" and python_version < "4.0"
pydantic-settings==2.7.1 ; python_version >= "3.11" and python_version < "4.0"
pydantic==2.10.5 ; python_version >= "3.11" and python_version < "4.0"
pyparsing==3.2.1 ; python_version >= "3.11" and python_version < "4.0"
pypdf==5.1.0 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.11" and python_version < "4.0"
python-engineio==4.11.2 ; python_version >= "3.11" and python_version < "4.0"
python-http-client==3.3.7 ; python_version >= "3.11" and python_version < "4.0"
python-socketio==5.12.1 ; python_version >= "3.11" and python_version < "4.0"
pytz==2024.2 ; python_version >= "3.11" and python_version < "4.0"
quill-delta[html]==1.0.3 ; python_version >= "3.11" and python_version < "4.0"
redis==5.2.1 ; python_version >= "3.11" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.11" and python_version < "4.0"
rq==2.1.0 ; python_version >= "3.11" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.11" and python_version < "4"
sendgrid==7.0.0rc2 ; python_version >= "3.11" and python_version < "4.0"
simple-websocket==1.1.0 ; python_version >= "3.11" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
sqlalchemy-utils==0.41.2 ; python_version >= "3.11" and python_version < "4.0"
sqlalchemy==2.0.37 ; python_version >= "3.11" and python_version < "4.0"
starkbank-ecdsa==2.2.0 ; python_version >= "3.11" and python_version < "4.0"
tqdm==4.67.1 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.11" and python_version < "4.0"
unidecode==1.3.8 ; python_version >= "3.11" and python_version < "4.0"
uritemplate==4.1.1 ; python_version >= "3.11" and python_version < "4.0"
urllib3==2.3.0 ; python_version >= "3.11" and python_version < "4.0"
uvicorn==0.34.0 ; python_version >= "3.11" and python_version < "4.0"
werkzeug==3.1.3 ; python_version >= "3.11" and python_version < "4.0"
wsproto==1.2.0 ; python_version >= "3.11" and python_version < "4.0"
wtforms==3.2.1 ; python_version >= "3.11" and python_version < "4.0"
